+++ CHANGELOG.md
@@ -1,2 +1,12 @@
+## [1.1.0] (Feb 24, 2025)
+### Features:
+- `onCompleted` and `onError` callbacks when call `execuute()`
+- `validate` method for validation purpose
+### Hotfix:
+- Notify when state changed to `running`
+- Check if command is running and not run until command completes
+- Typography
+- ErrorWrapper handler
+
 ## [1.0.0] (Feb 18, 2025)
 - Initial release